update home content_item code.

+++ lib/theme/theme_dimens.dart
@@ -5,27 +5,27 @@
 /// @description
 class ThemeDimens {
   // 全局通用组件数值
-  static const int toolbar_height = 56;
-  static const int list_state_item_height = 60;
+  static const double toolbar_height = 56;
+  static const double list_state_item_height = 60;
 
 // 全局通用radius值
-  static const int offset_radius_large = 32;
-  static const int offset_radius_medium = 16;
-  static const int offset_radius_small = 8;
+  static const double offset_radius_large = 32;
+  static const double offset_radius_medium = 16;
+  static const double offset_radius_small = 8;
 
 // 全局通用样式 （统一全局margin padding radius值）
-  static const int offset_large = 16;
-  static const int offset_medium = 8;
-  static const int offset_small = 4;
+  static const double offset_large = 16;
+  static const double offset_medium = 8;
+  static const double offset_small = 4;
 
 // 全局通用字体size值
-  static const int font_size_large_xx = 20;
-  static const int font_size_large_x = 18;
-  static const int font_size_large = 16;
-  static const int font_size_medium = 14;
-  static const int font_size_small = 12;
-  static const int font_size_small_x = 10;
+  static const double font_size_large_xx = 20;
+  static const double font_size_large_x = 18;
+  static const double font_size_large = 16;
+  static const double font_size_medium = 14;
+  static const double font_size_small = 12;
+  static const double font_size_small_x = 10;
 
 // 通用圆角线宽度
-  static const int border_width = 2;
+  static const double border_width = 2;
 }
